TDL-592: Creating a different file for testing a new support added in the tap
